refactor: standardize ErrorBlock and InfoBlock
