+++ frontend/src/components/common/index.ts
@@ -8,7 +8,7 @@ export { default as Pagination } from './pagination'
 export { default as TextArea } from './text-area'
 export { default as FileInput } from './file-input'
 export { default as DetailBlock } from './detail-block'
-export { default as PrimaryInfoBlock } from './primary-info-block'
+export { default as InfoBlock } from './info-block'
 export { default as ErrorBlock } from './error-block'
 export { default as PreviewBlock } from './preview-block'
 export { default as Modal } from './modal'
