+++ frontend/src/components/common/info-block/index.ts
@@ -1,0 +1,1 @@
+export { default } from './InfoBlock'
