+++ frontend/src/components/common/message-block/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MessageBlock'
