--- frontend/src/components/common/primary-info-block/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PrimaryInfoBlock'
