+++ frontend/src/components/dashboard/create/sms/SMSRecipients.tsx
@@ -18,7 +18,7 @@ import {
   TextButton,
   StepHeader,
   StepSection,
-  PrimaryInfoBlock,
+  InfoBlock,
 } from 'components/common'
 import { SMSCampaign, SMSPreview } from 'classes'
 import { sendTiming } from 'services/ga.service'
@@ -156,13 +156,12 @@ const SMSRecipients = ({
           />
         </CsvUpload>
         {isDemo && (
-          <PrimaryInfoBlock>
-            <h4>Limited to 20 recipients</h4>
+          <InfoBlock title="Limited to 20 recipients">
             <span>
               You can only send out to 20 recipients per test campaign. Only the
               first 20 rows in your CSV file will be taken.
             </span>
-          </PrimaryInfoBlock>
+          </InfoBlock>
         )}
         <ErrorBlock>{errorMessage}</ErrorBlock>
       </StepSection>
